refactor: create sass file directory
