--- frontend/src/index.scss
@@ -1,7 +1,0 @@
-@use './styles/variables';
-
-* {
-  padding: 0;
-  margin: 0;
-  font-family: variables.$general-typography;
-}
