+++ frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 // import App from './App'
-import './index.scss'
+import './scss/index.scss'
 import LogIn from './views/LogIn/LogIn'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
