+++ frontend/src/scss/abstracts/_variables.scss
@@ -1,0 +1,10 @@
+@import url('https://fonts.googleapis.com/css2?family=Manrope:wght@200;300;400&family=Mansalva&display=swap');
+
+:root {
+  /*fonts*/
+  --general-typography: 'Manrope', sans-serif;
+  --title-typography: 'Mansalva', cursive;
+
+  /*colors*/
+  --primary: #ff3366;
+}
