+++ frontend/src/scss/index.scss
@@ -1,0 +1,10 @@
+/*abstracts*/
+@use './abstracts/variables';
+
+/*base*/
+@use './base/base';
+@use './base/resets';
+
+/*views*/
+
+/*components*/
