--- frontend/src/styles/_variables.scss
@@ -1,6 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Manrope:wght@200;300;400&family=Mansalva&display=swap');
-
-$general-typography: 'Manrope', sans-serif;
-$title-typography: 'Mansalva', cursive;
-
-$primary: #ff3366;
